refactor: import external test configuration for clock and cleanup code
