+++ src/main/java/tech/gruppone/stalker/server/repositories/UserDataRepository.java
@@ -1,7 +1,5 @@
 package tech.gruppone.stalker.server.repositories;
 
-import java.time.LocalDate;
-import org.springframework.data.r2dbc.repository.Query;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import tech.gruppone.stalker.server.model.db.UserDataDao;
